+++ src/examples/Modules/Example.Schemas/Dtos/AdditionalData.cs
@@ -1,9 +1,10 @@
-﻿using System;
+﻿using System.Collections.Generic;
 
 namespace Example.Schemas.Dtos;
 
 public class AdditionalData
 {
-    public DateTime DateTime { get; set; }
+    public List<string> Items { get; set; }
     public string Guid { get; set; }
+    public string Name { get; set; }
 }