+++ src/examples/Server/Example.Server.Console/Program.cs
@@ -4,12 +4,14 @@ using BIDTP.Dotnet.Core.Iteraction.Enums;
 using BIDTP.Dotnet.Core.Iteraction.Options;
 using BIDTP.Dotnet.Core.Iteraction.Providers;
 using Example.Server.Core.Workers;
+using Example.Server.Domain.Auth.Middlewares;
 using Example.Server.Domain.Auth.Providers;
 using Example.Server.Domain.Colors.Controllers;
 using Example.Server.Domain.Colors.Providers;
 using Example.Server.Domain.Elements.Controllers;
 using Example.Server.Domain.Elements.Repositories;
 using Example.Server.Domain.Messages.Controllers;
+using Example.Server.Domain.Messages.Middlewares;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
@@ -37,7 +39,8 @@ namespace Example.Server.Console
              
             builder.AddRoute("PrintMessage", JustChickenGuard, ColorController.GetRandomColor);
             builder.AddRoute("GetElements", ElementController.GetElements);
-            builder.AddRoute("GetMappedObjectFromObjectContainer", SendMessageController.GetMappedObjectWithMetadataFromObjectContainer);
+            builder.AddRoute("GetMappedObjectFromObjectContainer", ObjectContainerMiddleware.Handle,
+                SendMessageController.GetMappedObjectWithMetadataFromObjectContainer);
             
             Task JustChickenGuard(Context context)
             {